[REFACTOR] 강사 프로젝트 조회 수정
