+++ frontend/src/views/project/ProjectDetailView.vue
@@ -13,6 +13,7 @@ const router = useRouter();
 
 // 프로젝트 ID 가져오기
 const projectSeq = route.params.id;
+const projectTeamList = ref([]);
 
 // API 호출 함수
 const fetchProjectDetail = async () => {
@@ -23,6 +24,19 @@ const fetchProjectDetail = async () => {
       }
     });
     project.value = projectResponse.data.data;
+    console.log("project : ", project.value)
+
+    try {
+      const projectTeamResponse = await axios.get(`http://localhost:8086/api/v1/team`, {
+        headers: {
+          'Authorization': `Bearer ${localStorage.getItem('authToken')}`
+        }
+      })
+      projectTeamList.value = projectTeamResponse.data.data;
+      console.log(projectTeamList.value);
+    } catch (error) {
+      console.error('팀 정보를 불러오는 중 에러가 발생했습니다.', error);
+    }
   } catch (error) {
     console.error('프로젝트 정보를 불러오는 중 에러가 발생했습니다:', error);
   }
@@ -33,18 +47,37 @@ onMounted(fetchProjectDetail);
 </script>
 
 <template>
-  <template v-if="project">
+  <template v-if="projectTeamList && project">
       <div class="row card-container" v-if="project.projectMentorYn=== 'Y'">
         <div class="col-md-6">
-          <MentorProjectDetail :project="project" class="card"/>
+          <b-card class="h-100">
+            <MentorProjectDetail v-if="projectTeamList.length > 0" :project="project" :project-team-list="projectTeamList" class="card"/>
+
+            <MentorProjectTeamList v-if="projectTeamList > 0" :project-team-list="projectTeamList"/>
+            <div class="no-team-container text-center" v-else>
+              <!-- 경고 아이콘 -->
+              <div class="warning-icon">
+                <i class="bi bi-exclamation-triangle-fill"></i>
+              </div>
+              <!-- 텍스트 -->
+              <p class="mt-3">
+                진행 중인 팀이 존재하지 않습니다.
+              </p>
+              <p>
+                T-Building에서 제공하는 자동 팀 빌딩모드로<br />
+                지금 바로 팀을 빌딩해 보세요!
+              </p>
+              <b-button variant="dark" style="float: right; height: 35px;">팀 빌딩하기</b-button>
+            </div>
+          </b-card>
         </div>
-        <TeamDetail :team-seq="project.teamSeq" class="card" />
+<!--        <TeamDetail v-if="projectTeamList.length > 1" :team-seq="project.teamSeq" class="card" />-->
       </div>
       <div class="row card-container" v-else>
         <MemberProjectDetail :project="project" class="card" />
-        <TeamDetail :team-seq="project.teamSeq" class="card" />
+        <TeamDetail v-if="projectTeamList.length === 1" :team-seq="project.teamSeq" class="card" />
       </div>
-    </template>
+  </template>
     <div v-else>
       <p>Loading...</p>
     </div>
@@ -64,14 +97,23 @@ onMounted(fetchProjectDetail);
   flex-direction: column; /* 카드 안의 요소들을 수직으로 배치 */
 }
 
-.main-container {
+.no-team-container {
   display: flex;
-  padding: 20px;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 50vh; /* 화면 중앙에 배치 */
+  text-align: center;
 }
 
-/* 좌측 프로젝트 정보 */
-.project-info {
-  width: 50%;
-  padding-right: 20px;
+.warning-icon {
+  font-size: 50px;
+  color: #f0ad4e; /* 경고 색상 */
 }
+
+
+.team-button {
+  margin-left: auto;
+}
+
 </style>
